+++ examples/mqtt_client.py
@@ -1,3 +1,14 @@
+"""
+:Example:
+
+    >>> tiny-listener mqtt_client:app
+    Log handler: 001 => bytearray(b'LOG info: ...')
+    Log handler: 002 => bytearray(b'LOG error: ...')
+    Power handler: 001 => bytearray(b'POWER 20%')
+    Power handler: 002 => bytearray(b'POWER 30%')
+"""
+
+
 from hbmqtt.client import MQTTClient
 from hbmqtt.mqtt.constants import QOS_0
 from hbmqtt.mqtt.publish import PublishPacket
@@ -42,5 +53,3 @@ async def fn(event: Event):
     await client.publish('/sys/device/002/log', b'LOG error: ...')
     await client.publish('/sys/device/001/power', b'POWER 20%')
     await client.publish('/sys/device/002/power', b'POWER 30%')
-
-app.run()