+++ tiny_listener/context.py
@@ -102,11 +102,10 @@ class Context(metaclass=__UniqueCTX):
     def todo(
             self,
             name: str,
-            block: bool = False,
             timeout: Optional[float] = None,
             data: Optional[Dict] = None
     ) -> Coroutine or None:
-        return self.listener.todo(name=name, cid=self.cid, block=block, timeout=timeout, data=data)
+        return self.listener.todo(name=name, cid=self.cid, timeout=timeout, data=data)
 
     @classmethod
     def exist(cls, cid: str) -> bool: